feat(execution): add plasma deploy config and read fee setter from roles
Add the plasma network (chain id 9745) to hardhat config with Routescan
verification, define the executors to deploy (uniswap_v3, fluid_v1,
vm:curve, native_wrapper) with their constructor args, and add plasma
role assignments.

Replace the ROUTER_FEE_SETTER env var with a lookup in roles.json so
deploy-fee-calculator.js and deploy-router.js share one source of truth
via resolveRolesNetwork, now exported from utils.js.
