+++ crates/tycho-execution/config/executor_deployments.json
@@ -182,5 +182,23 @@
       "contract": "MetricExecutor",
       "args": ["0x681E908B8ab57c49C74D770F369754CCC3E1ae09"]
     }
+  },
+  "plasma": {
+    "uniswap_v3": { "contract": "UniswapV3Executor", "args": [] },
+    "fluid_v1": {
+      "contract": "FluidV1Executor",
+      "args": ["0x52Aa899454998Be5b000Ad077a46Bbe360F4e497"]
+    },
+    "vm:curve": {
+      "contract": "CurveExecutor",
+      "args": [
+        "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE",
+        "0x0000000000000000000000000000000000000000"
+      ]
+    },
+    "native_wrapper": {
+      "contract": "NativeWrapExecutor",
+      "args": ["0x6100E367285b01F48D07953803A2d8dCA5D19873"]
+    }
   }
 }
